Code/backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
faiss-cpu
requests
tqdm
sentence-transformers
openai
python-dotenv
google-genai
selenium>=4.21.0
webdriver-manager>=4.0.2
beautifulsoup4>=4.12.3
lxml>=5.3.0
torch
transformers